dom/credentialmanagement/identity/tests/mochitest/test_continue_on.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Happypath Test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    setupTest("continue_on").then(
      function () {
        SpecialPowers.wrap(document).notifyUserGestureActivation();
        return navigator.credentials.get({
          identity: {
            mode: "active",
            providers: [{
              configURL: "https://example.net/tests/dom/credentialmanagement/identity/tests/mochitest/server_manifest.sjs",
              clientId: "mochitest",
              nonce: "nonce"
            }]
          }
        });
      }
    ).then((cred) => {
      ok(true, "successfully got a credential");
      is(cred.token,
        "result_from_inside_continue",
        "Correct token on the credential.");
      is(cred.id,
        "1234",
        "Correct id on the credential");
      is(cred.type,
        "identity",
        "Correct type on the credential");
    }).catch(() => {
      ok(false, "must not have an error");
    }).finally(() => {
      async function testStorageAccessPermissionMissing() {
        ok(await SpecialPowers.testPermission('3rdPartyStorage^https://example.net', SpecialPowers.Ci.nsIPermissionManager.UNKNOWN_ACTION, document), "The popup interacted and should not have given a permission");
      }
      return testStorageAccessPermissionMissing();
    }).then(() => {
      SimpleTest.finish();
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">This is the main happypath test. We get a credential in a way that should work. This includes simplifying some logic like exactly one account and provider.</div>
<pre id="test"></pre>
</body>
</html>
